+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Content moved to Workspace and Content removed
 * Added library of robe-ajax and ajaxsetup settings
 * was update Permission style
+* was update Header style